app/requirements.txt
streamlit==1.40.1
streamlit-audiorecorder==0.0.6
pydub==0.25.1
webrtcvad
openai
requests
openai-whisper